+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 Version history for `cardano-ledger-release-tool`
 
+## 0.4.0.1
+
+*
+
 ## 0.4.0.0
 
 * Add new `cabal` subcommands: `list-bins`, `run`, `test`